feat: architecture improvement
