+++ frontend/src/components/Button/index.tsx
@@ -1,0 +1,2 @@
+export { ButtonCard } from './ButtonCard';
+export { ButtonRow } from './ButtonRow';
