+++ frontend/src/components/Navbar/Navbar.css
@@ -1,93 +1,93 @@
 nav {
-  display: grid;
-  grid-template-columns: min-content min-content auto;
-  align-items: center;
-  padding: 1rem 0;
-  column-gap: 1rem;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.15);
-  user-select: none;
+    display: grid;
+    grid-template-columns: min-content max-content max-content;
+    align-items: center;
+    padding: 1rem 0;
+    column-gap: 1rem;
+    border-bottom: 1px solid rgba(0, 0, 0, 0.15);
+    user-select: none;
 }
 
 nav img {
-  display: flex;
-  align-items: center;
-  height: 24px;
+    display: flex;
+    align-items: center;
+    height: 24px;
 }
 
-nav .links { 
-  display: flex;
-  column-gap: 1rem;
+nav .links {
+    display: flex;
+    column-gap: 1rem;
 }
 
 nav .links.right {
-  display: flex;
-  align-items: center;
-  column-gap: 1rem;
-  margin-left: auto;
+    display: flex;
+    align-items: center;
+    column-gap: 1rem;
+    margin-left: auto;
 }
 
 nav .links.right a {
-  color: rgba(0, 0, 0, 0.45);
+    color: rgba(0, 0, 0, 0.45);
 }
 
 nav .links.right a:hover,
 nav .links.right a.active {
-  color: #000;
+    color: #000;
 }
 
 nav .hamburger {
-  display: none;
-  cursor: pointer;
-  width: 14px;
+    display: none;
+    cursor: pointer;
+    width: 14px;
 }
 
 nav .hamburger .line {
-  width: 14px;
-  height: 1px;
-  background-color: rgba(0, 0, 0, 0.25);
-  border-radius: 0.25rem;
+    width: 14px;
+    height: 1px;
+    background-color: rgba(0, 0, 0, 0.25);
+    border-radius: 0.25rem;
 }
 
 nav .hamburger.active .line {
-  background-color: #000;
+    background-color: #000;
 }
 
 @media (max-width: 768px) {
-  nav {
-    grid-template-columns: auto 14px;
-    row-gap: 1rem;
-  }
+    nav {
+        grid-template-columns: auto 14px;
+        row-gap: 1rem;
+    }
 
-  nav .links,
-  nav .links.right {
-    display: none;
-  }
+    nav .links,
+    nav .links.right {
+        display: none;
+    }
 
-  nav button,
-  nav input {
-    width: 100%;
-  }
+    nav button,
+    nav input {
+        width: 100%;
+    }
 
-  nav .hamburger {
-    display: grid;
-    grid-template-rows: repeat(3, 1fr);
-    row-gap: 0.15rem;
-  }
+    nav .hamburger {
+        display: grid;
+        grid-template-rows: repeat(3, 1fr);
+        row-gap: 0.15rem;
+    }
 
-  nav.collapsed .links,
-  nav.collapsed .links.right {
-    display: grid;
-    grid-column: 1 / span 2;
-    padding: 0;
-    gap: 10px;
-    width: 100%;
-  }
+    nav.collapsed .links,
+    nav.collapsed .links.right {
+        display: grid;
+        grid-column: 1 / span 2;
+        padding: 0;
+        gap: 10px;
+        width: 100%;
+    }
 
-  nav.collapsed .links {
-    padding: 0.5rem 0;
-  }
+    nav.collapsed .links {
+        padding: 0.5rem 0;
+    }
 
-  nav .links.right {
-    grid-template-columns: auto;
-  }
+    nav .links.right {
+        grid-template-columns: auto;
+    }
 }
