+++ frontend/src/index.css
@@ -1,12 +1,25 @@
+@import url('https://fonts.googleapis.com/css2?family=Rubik:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
+:root {
+    --charlieWhite: #f5f4f3;
+    --charlieGrey: #e5e5e5;
+    --charlieBlack: #000;
+    --charliePink: #ff04c9;
+    --charlieOrange: #ff6804;
+    --charlieGradient: linear-gradient(-37deg, #ff6804, #ff04c9);
+    --charlieGradientAlt: linear-gradient(-37deg, #5243f9, #43f9e1);
+}
+
 * {
     box-sizing: border-box;
     margin: 0;
     padding: 0;
-    font-family: 'Inter';
+    font-family: 'Rubik', sans-serif;
 }
 
 html,
 body {
+    background: var(--charlieWhite);
     color: rgba(0, 0, 0, 0.87);
     height: 100%;
     width: 100vw;
@@ -30,26 +43,26 @@ button {
     outline: none;
     cursor: pointer;
     font-weight: 700;
-    padding: 1rem;
+    padding: 0.65rem .75rem;
     transition: all .15s ease-in-out;
 }
 
 button.primary {
-    background: linear-gradient(-37deg, #ff6804, #ff04c9);
+    background: var(--charlieGradient);
     position: relative;
     background-position: 0 0;
 }
 
 button.primary .content {
     position: relative;
-    color: #fff;
+    color: var(--charlieWhite);
     z-index: 5;
     height: initial;
 }
 
 button.primary:hover,
 button.primary:focus {
-    background: linear-gradient(-37deg, #ff6804, #ff04c9);
+    background: var(--charlieGradient);
 }
 
 button.primary::before {
@@ -73,7 +86,7 @@ button.primary::after {
     left: -2px;
     width: calc(100% + 4px);
     height: calc(100% + 4px);
-    background: linear-gradient(-37deg, #ff6804, #ff04c9);
+    background: var(--charlieGradient);
     z-index: -1;
     filter: blur(10px);
     opacity: .25;
@@ -91,17 +104,17 @@ button.primary:focus::before {
 }
 
 button.primary.secondary {
-    background: linear-gradient(-37deg, #5243f9, #43f9e1);
+    background: var(--charlieGradientAlt);
     color: #fff;
 }
 
 button.primary.secondary:hover,
 button.primary.secondary:focus {
-    background: linear-gradient(-37deg, #5243f9, #43f9e1);
+    background: var(--charlieGradientAlt);
 }
 
 button.primary.secondary::after {
-    background: linear-gradient(-37deg, #5243f9, #43f9e1);
+    background: var(--charlieGradientAlt);
 }
 
 button:disabled {
@@ -143,23 +156,14 @@ input {
 }
 
 .card {
-    border: 1px solid rgba(0, 0, 0, 0.15);
+    background: #f5f4f3;
     border-radius: 3px;
-    margin-bottom: 1rem;
-    margin: 1rem;
-}
-
-.card .card-header {
-    border-bottom: 1px solid rgba(0, 0, 0, 0.15);
     padding: 1rem;
-}
-
-.card .card-body {
-    padding: 1rem;
+    margin: 1rem 0;
 }
 
 .gradient {
-    background: linear-gradient(-37deg, #ff6804, #ff04c9);
+    background: var(--charlieGradient);
     -webkit-background-clip: text;
     -webkit-text-fill-color: transparent;
     position: relative;
@@ -173,15 +177,14 @@ input {
 }
 
 .gradient::before {
-    content: 'one place';
+    content: 'one click';
     position: absolute;
     top: 0;
     left: 0;
     width: 100%;
     height: 100%;
-    background: linear-gradient(-37deg, #ff6804, #ff04c9);
+    background: var(--charlieGradient);
     z-index: -1;
-
     background-blend-mode: screen;
     -webkit-background-clip: text;
 }
@@ -193,7 +196,7 @@ input {
     left: -2px;
     width: calc(100% + 4px);
     height: calc(100% + 4px);
-    background: linear-gradient(-37deg, #ff6804, #ff04c9);
+    background: var(--charlieGradient);
     z-index: -1;
     filter: blur(50px);
     opacity: .35;
