+++ frontend/src/pages/Account.tsx
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router-dom'
 
 import { useAuthentication } from '../hooks'
 
-import { ButtonForm, Buttons } from './'
+import { Button, ButtonForm, Buttons } from './'
 
 const Account = () => {
     const {
@@ -22,6 +22,7 @@ const Account = () => {
     return <Routes>
         <Route path="/" element={<Buttons />} />
         <Route path="/buttons/new/" element={<ButtonForm />} />
+        <Route path="/buttons/:buttonId/" element={<Button />} />
         <Route path="/buttons/:buttonId/edit/" element={<ButtonForm isEdit={true} />} />
     </Routes>
 }
