+++ frontend/src/pages/Button.tsx
@@ -1,23 +1,9 @@
-import { useEffect, useState } from "react";
-
-import { useParams } from "react-router-dom";
+import { Link, useParams } from "react-router-dom"
 
 import { useQuery } from "@tanstack/react-query";
 
-import { TokenRow } from "../components";
-
-import { useColor } from "../hooks";
-
-import { getBalances } from "../utils";
-
-import "./Button.css";
-
 const Button = () => {
-    const { buttonId } = useParams<{ buttonId: string }>();
-
-    const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-
-    const [balances, setBalances] = useState<any>(null);
+    const { buttonId } = useParams();
 
     const {
         isLoading,
@@ -38,72 +24,27 @@ const Button = () => {
             }).then((res) => res.json()),
     });
 
-    const textColor = useColor(data?.hex_color);
-
-    useEffect(() => {
-        if (!data) return;
-
-        getBalances({
-            address: "0x62180042606624f02d8a130da8a3171e9b33894d",
-            tokens: data.tokens,
-            includeZeros: true
-        }).then(({ results }) => {
-            setBalances(results)
-        })
-    }, [isLoading])
-
     if (isLoading) return <>{"Loading..."}</>;
 
     if (error) return <>{"An error has occurred: " + error.message}</>;
 
-    return (
-        <>
-            <div
-                className={isModalOpen ? "modal" : "modal hidden"}
-                onClick={() => setIsModalOpen(false)}>
-                <div className="modal-content">
-                    <span className="close">
-                        &times;
-                    </span>
+    return <>
+        <Link to="/account/" children={
+            <button>Back</button>
+        } />
 
-                    <h1>{data.text}</h1>
+        <Link to={`/account/buttons/${buttonId}/edit`} children={
+            <button>Edit</button>
+        } />
 
-                    {data.tokens
-                        .sort((a: any, b: any) => b.chain_id - a.chain_id)
-                        .map((token: any) => {
-                            const balance = balances?.find((balance: any) => balance.address === token.address)?.balance
+        <Link to={`/hosted/buttons/${buttonId}/embed`} children={
+            <button>Preview</button>
+        } />
 
-                            const chainId = token.chain_id;
-                            const previousChainId = data.tokens[data.tokens.indexOf(token) - 1]?.chain_id;
+        <h1>{data.text}</h1>
 
-                            return (
-                                <TokenRow
-                                    key={`${token.ethereum_address}-${token.chain_id}`}
-                                    token={token}
-                                    balance={balance}
-                                    first={chainId !== previousChainId}
-                                />
-                            )
-                        })}
+        <p>{data.description}</p>
+    </>
+}
 
-                    <button className="delegate">Delegate now</button>
-
-                    <p>Powered by <strong>Charlie</strong>.</p>
-                </div>
-            </div >
-
-            <button
-                className="delegation-button"
-                style={{
-                    backgroundColor: data.hex_color,
-                    color: textColor,
-                }}
-                onClick={() => setIsModalOpen(!isModalOpen)}
-            >
-                {data.text}
-            </button>
-        </>
-    );
-};
-
-export { Button };
+export { Button }
