+++ frontend/src/pages/index.tsx
@@ -3,6 +3,7 @@ export { Static } from './Static'
 
 export { Account } from './Account'
 export { Button } from './Button'
+export { ButtonEmbed } from './ButtonEmbed'
 export { ButtonForm } from './ButtonForm'
 export { Buttons } from './Buttons'
 export { Home } from './Home'
